+++ index.d.ts
@@ -3,4 +3,4 @@ import Sidebar, { moduleNow, tabAdmin, domSidebar } from './src/index.vue.js';
 export { Sidebar, moduleNow, tabAdmin, domSidebar };
 
 
-export { TabAdmin, Tab } from './src/tab-admin.js';
+export { default as TabAdmin, Tab } from './src/tab-admin.js';